syntax.md -- remove space in "{ : .note}"
This page has the command "{: .note}" in two places. I have no idea what that is supposed to do, but in one case, it has a space before the colon, and shows up in the formatted output, and in the other it does not have the space before the colon, and is not visible in the formatted output.

I am *guessing* that the space is a typo and removing it. But I still don't know what "{: .note}" is supposed to be doing, so this may be a mistake.

Signed-off-by: Stavros Macrakis <134456002+smacrakis@users.noreply.github.com>